# OpenAPI.yaml
openapi: "3.0.3"
info:
  title: Bangumi Recorder API
  version: "2.0.0"
  description: |
    自托管的番剧追番记录器 API v2。

    所有 v2 接口统一返回格式：
    ```json
    { "status": 0, "data": ..., "message": "" }
    ```
    - `status`: `0` 成功 / `-1` 报错
    - `data`: 所有响应数据
    - `message`: 错误描述（成功时为空）

    错误时返回适当的 HTTP 状态码（400 / 404 / 409 / 401 / 500 等）。
servers:
  - url: ""
    description: 与前端同源
paths:
...
  /api/v2/tokens/{id}:
    put:
      summary: 更新 API Token
      tags: [API Token 管理]
      security: [{ BearerJWT: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                permissions: { type: integer }
                is_active: { type: boolean }
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiOk"
        "404":
          description: Token 未找到
    delete:
      summary: 删除/吊销 API Token
      tags: [API Token 管理]
      security: [{ BearerJWT: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiOk"
        "404":
          description: Token 未找到
  /api/v2/tokens/permissions:
    get:
      summary: 获取权限标签列表
      tags: [API Token 管理]
      security: [{ BearerJWT: [] }]
      responses:
        "200":
          description: 权限标签列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PermissionLabel"

  # ────────────────────────────── 搜索 ──────────────────────────────
...
components:
  securitySchemes:
    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 在 /api/v2/auth/login 获取 Token，以 `Bearer <token>` 形式传入 Authorization 头。
  schemas:
...
    ApiOk:
      type: object
      properties:
        status: { type: integer, enum: [0] }
        data: { nullable: true, default: null }
        message: { type: string, nullable: true }
...
    PermissionLabel:
      type: object
      properties:
        label: { type: string, description: "权限名称" }
        value: { type: integer, description: "权限 bit 值" }
        description: { type: string, description: "权限说明" }
      required: [label, value, description]

    # ── 版本 ──
...